+++ app/Helpers/helpers.php
@@ -155,11 +155,107 @@ if (! function_exists('responseJSON')) {
 }
 
 if (! function_exists('getPrice')) {
-    function getPrice($num)
+    function getPrice($num, $class = null)
     {
-        return '<div class="d-flex">
+        return '<div class="d-flex '.$class.'">
                     <span class="me-2">Rp. </span>
                     <span>'.number_format($num, 0, ',', '.').'</span>
                 </div>';
+    }
+}
+
+if (! function_exists('getWorkingDay')) {
+    function getWorkingDay($now = null)
+    {
+        // do strtotime calculations just once
+        $endDate = strtotime(date('Y-m-t'));
+        $startDate = strtotime(date('Y-m-01'));
+
+        // jika memiliki tanggal
+        if ($now) {
+            $startDate = strtotime($now);
+        }
+
+        //The total number of days between the two dates. We compute the no. of seconds and divide it to 60*60*24
+        //We add one to inlude both dates in the interval.
+        $days = ($endDate - $startDate) / 86400 + 1;
+
+        $no_full_weeks = floor($days / 7);
+        $no_remaining_days = fmod($days, 7);
+
+        //It will return 1 if it's Monday,.. ,7 for Sunday
+        $the_first_day_of_week = date("N", $startDate);
+        $the_last_day_of_week = date("N", $endDate);
+
+        //---->The two can be equal in leap years when february has 29 days, the equal sign is added here
+        //In the first case the whole interval is within a week, in the second case the interval falls in two weeks.
+        if ($the_first_day_of_week <= $the_last_day_of_week) {
+            if ($the_first_day_of_week <= 6 && 6 <= $the_last_day_of_week) $no_remaining_days--;
+            if ($the_first_day_of_week <= 7 && 7 <= $the_last_day_of_week) $no_remaining_days--;
+        }
+        else {
+            // (edit by Tokes to fix an edge case where the start day was a Sunday
+            // and the end day was NOT a Saturday)
+
+            // the day of the week for start is later than the day of the week for end
+            if ($the_first_day_of_week == 7) {
+                // if the start date is a Sunday, then we definitely subtract 1 day
+                $no_remaining_days--;
+
+                if ($the_last_day_of_week == 6) {
+                    // if the end date is a Saturday, then we subtract another day
+                    $no_remaining_days--;
+                }
+            }
+            else {
+                // the start date was a Saturday (or earlier), and the end date was (Mon..Fri)
+                // so we skip an entire weekend and subtract 2 days
+                $no_remaining_days -= 2;
+            }
+        }
+
+        //The no. of business days is: (number of weeks between the two dates) * (5 working days) + the remainder
+//---->february in none leap years gave a remainder of 0 but still calculated weekends between first and last day, this is one way to fix it
+        $workingDays = $no_full_weeks * 5;
+        if ($no_remaining_days > 0 )
+        {
+            $workingDays += $no_remaining_days;
+        }
+
+        $holidays = [
+            '2022-01-01',
+            '2022-01-01', // Tahun Baru Masehi 2022
+            '2022-01-02', // Tahun Baru Imlek
+            '2022-01-03', // Isra' Mi'raj
+            '2022-01-04', // Nyepi
+            '2022-01-05', // Wafat Isa Al-masih
+            '2022-01-06', // Hari Buruh
+            '2022-01-07', // Idul Fitri
+            '2022-01-08', // Idul Fitri
+            '2022-01-09', // Libur Perusahaan
+            '2022-01-10', // Cuti Bersama
+            '2022-01-11', // Cuti Bersama
+            '2022-01-12', // Waisak
+            '2022-01-13', // Kenaikan Isa
+            '2022-01-14', // Pancasila
+            '2022-01-15', // Idul Adha
+            '2022-01-16', // Tahun Baru Hijriah
+            '2022-01-17', // Kemerdekaan
+            '2022-01-18', // Maulid Nabi
+            '2022-01-19', //	Natal
+            '2022-01-20', //	Libur Perusahaan
+            '2022-01-21', //	Cuti Bersama
+            '2022-01-22', //	Cuti Bersama
+        ];
+
+        //We subtract the holidays
+        foreach($holidays as $holiday){
+            $time_stamp=strtotime($holiday);
+            //If the holiday doesn't fall in weekend
+            if ($startDate <= $time_stamp && $time_stamp <= $endDate && date("N",$time_stamp) != 6 && date("N",$time_stamp) != 7)
+                $workingDays--;
+        }
+
+        return $workingDays;
     }
 }