+++ app/Models/User.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Auth\User as Authenticatable;
@@ -63,5 +64,49 @@ class User extends Authenticatable
             session('theme', 'light');
         }
         return $theme;
+    }
+
+    public function getSetRepairAttribute()
+    {
+        return Job::where('status', Config::all()->first()->invoice_job_status_invoice)
+            ->whereYear('collection_at', Carbon::now()->year)
+            ->whereMonth('collection_at', Carbon::now()->month)
+            ->get()
+            ->count();
+    }
+
+    public function getSpeedRepairAttribute()
+    {
+        $jobs =  Job::where('status', Config::all()->first()->invoice_job_status_invoice)
+            ->whereYear('collection_at', Carbon::now()->year)
+            ->whereMonth('collection_at', Carbon::now()->month)
+            ->get();
+
+        $count = 0;
+        foreach ($jobs as $row) {
+            if ($row->day <= 2) {
+                $count += 1;
+            }
+        }
+
+        return $count;
+    }
+
+    public function getIncomeAttribute()
+    {
+        $jobs =  Job::where('status', Config::all()->first()->invoice_job_status_invoice)
+            ->whereYear('collection_at', Carbon::now()->year)
+            ->whereMonth('collection_at', Carbon::now()->month)
+            ->get();
+
+        $sum = 0;
+        foreach ($jobs as $row) {
+            $total = Invoice::where('job', $row->id)->get()->first()->total_no_disc;
+            if ($total > 0) {
+                $sum += $total;
+            }
+        }
+
+        return $total;
     }
 }