+++ resources/views/layout/job/table.blade.php
@@ -62,7 +62,13 @@
                 <tr>
                     <td class="ps-3 text-muted w-1-slot">{{ $table['column'] == 'id' && $table['sort'] == 'desc' ? $data->total() - ($data->firstItem() + $key) + 1 : $data->firstItem() + $key }}</td>
                     <td>
-                        <small class="d-inline-flex">{{ date('l, d/m/Y H:i', strtotime($row->created_at)) }}</small>
+                        <small class="d-inline-flex {{ $row->quality_report ? 'text-secondary' : '' }} {{ $row->dealer_report ? 'text-primary' : '' }}">{{ date('l, d/m/Y H:i', strtotime($row->created_at)) }}</small>
+                        @if($row->quality_report)
+                            <br><span class="{{ getBadge('secondary') }}">QC Pass</span>
+                        @endif
+                        @if($row->dealer_report)
+                            <br><span class="{{ getBadge('blue') }}">Dealer Checked</span>
+                        @endif
                     </td>
                     <td>
                         <span class="text-nowrap">
@@ -77,19 +83,18 @@
                     </td>
                     <td>
                         @if($row->invoice)
-                            <span class="text-nowrap">{{ $row->invoice }}</span>
+                            <a href="{{ url('invoice/'.$row->invoice.'/edit') }}" class="text-nowrap {{ $row->quality_report ? 'text-secondary' : '' }} {{ $row->dealer_report ? 'text-primary' : '' }}">
+                                <span>{{ $row->invoice_name }}</span>
+                            </a>
+                            @if($row->invoice_paid)
+                                <br><span class="{{ getBadge('primary') }}">Paid</span>
+                            @endif
                         @else
                             <span class="text-muted">-</span>
                         @endif
-                        @if($row->quality_report)
-                            <br><span class="{{ getBadge('purple') }}">Quality Checked</span>
-                        @endif
-                        @if($row->dealer_report)
-                            <br><span class="{{ getBadge('blue') }}">Dealer Checked</span>
-                        @endif
                     </td>
                     <td>
-                        <span>{{ $row->customer_name }}</span>
+                        <span class="{{ $row->quality_report ? 'text-secondary' : '' }} {{ $row->dealer_report ? 'text-primary' : '' }}">{{ $row->customer_name }}</span>
                         <br><small>{{ $row->phone }}</small>
                         @if($row->phone2)
                             <br><small class="text-muted">{{ $row->phone2 }}</small>
@@ -110,7 +115,7 @@
                         @endif
                     </td>
                     <td>
-                        <span>{{ $row->model }}</span>
+                        <span class="{{ $row->quality_report ? 'text-secondary' : '' }} {{ $row->dealer_report ? 'text-primary' : '' }}">{{ $row->model }}</span>
                         @if($row->serial)
                             <br><small class="text-muted">{{ $row->serial }}</small>
                         @endif
@@ -126,9 +131,9 @@
                     </td>
                     <td>
                         @if(strlen($row->service_info) > 30)
-                            <span class="text-nowrap" data-bs-toggle="tooltip" data-bs-placement="top" title="{{ $row->service_info }}">{{ substr_replace($row->service_info, '...', 30) }}</span>
+                            <span class="text-nowrap {{ $row->quality_report ? 'text-secondary' : '' }} {{ $row->dealer_report ? 'text-primary' : '' }}" data-bs-toggle="tooltip" data-bs-placement="top" title="{{ $row->service_info }}">{{ substr_replace($row->service_info, '...', 30) }}</span>
                         @else
-                            <span class="text-nowrap">{{ $row->service_info }}</span>
+                            <span class="text-nowrap {{ $row->quality_report ? 'text-secondary' : '' }} {{ $row->dealer_report ? 'text-primary' : '' }}">{{ $row->service_info }}</span>
                         @endif
                         @if(strlen($row->repair_info) > 30)
                             <br><small class="text-muted text-nowrap" data-bs-toggle="tooltip" data-bs-placement="top" title="{{ $row->repair_info }}">{{ substr_replace($row->repair_info, '...', 30) }}</small>